feat: add base layer for gturbine
I think this is generally correctly functioning, but I also think it
needs a little reorganization -- the erasure coding details all live in
the root of gturbine right now, but it should probably move to a package
dedicated to erasure coding. The final package organization won't be
totally clear until we get a little further with the other tooling to
allow integrating gturbine with other driver code.