feat: add gerasure package
This package enables using erasure coding to translate a slice of data
into erasure-coded shards that can be used to reconstitute the original
data. The first implementation is an off-the-shelf Reed-Solomon coding
library (which uses shards of the original data and separate parity
shards).

The interface should also enable rateless erasure coding, which shards
the data differently, should that use case come up.