feat: add gerasure package
This package enables using erasure coding to translate a slice of data
into erasure-coded shards that can be used to reconstitute the original
data. The first implementation is an off-the-shelf Reed-Solomon coding
library (which uses shards of the original data and separate parity
shards, typical for this form of erasure coding).

We also have compliance tests exercising a variety of data and parity
shard counts with varying original data sizes.

The gerasure interfaces should also enable rateless erasure coding,
which shards the data differently, should that use case come up.