feat: add gnetdag package
This adds a FixedTree type, which operates on just int values, making it
trivial to manage mapping parent-child relationships in a sorted slice.
The FixedTree nodes each have a fixed number of children.

This is intended to be used to map out parent-child relationships in a
tree for distributing block data over a public network, hence the
"g-net-dag" (Gordian, Network, Directed Acyclic Graph) name.