feat: initial sketch of BLS
This first implementation uses the minimized-signature form. I don't
think we have a use case for the minimized-key form, but the packages
are structured such that we can add it later.

I don't think this internal API is finalized yet; we will need to try it
out with the existing gcrypto signature proof interfaces to see what
refining it needs.

We may need some adjustments to account for the salt. I need to double
check whether we need any further custom data for the Domain Separation
Tag. Using the compressed forms of points is useful for serialization,
but we may want separate methods to expose the raw points, in order to
more efficiently aggregate keys and signatures.

There is an outstanding issue on the blst repository about compatibility
with go1.24, so the dependency on blst means that for now, using BLS
means sticking with go1.23.