feat: add sigtree package
The gmerkle tree implementation had some API issues that made it more
difficult than necessary to go back and forth between keys, signatures,
and sparse key IDs. So this sigtree package is meant to take the place
of gmerkle within gblsminsig.

It isn't complete yet, but it should work for prototyping, and the known
deficiencies can be addressed individually as they come up.

We are using a sigtree field in the minimized signature proof as of this
commit, but we are not yet interacting with it.