feat: add gblsminsig package
This uses the supranational/blst package for an implementation of BLS
signatures. We use the minimized signature form, hence the package name.

The package includes a BLS-backed gcrypto.PubKey implementation and a
CommonMessageSignatureProof implementation. There are likely some
missing details in the BLS setup, at least in adding configurability for
the signatures' salt. I think the domain separation tag is correct.

I cannot find the whitepaper or technical document that introduced me to
the concept, but the model of arranging the validators in a tree for key
and signature aggregation is based on the concept that the leftmost
validators are trustworthy and likely to be online, while the rightmost
validators are the least expected to vote. See further documentation on
the SignatureProof type.

It has not been fully integrated yet, but we have tests around the usage
of both of those types.

Next up are some slight interface changes that will affect the ed25519
implementation too. And I think we may need some special handling such
that the signature committed to chain is a "final aggregation" of keys
that would not normally be aggregated together. For instance, if there
are only four validators, then during consensus gossip, we would pair
0-1 and 2-3, but never 0-2 or 1-3. If the final votes were only from 0,
1, and 3 -- that is to say validator 2 was absent or perhaps voted nil
-- then we would commit with the aggregated signaure 0-1-3, to minimize
space used for signatures.

This also removes the gmerkle package, which was left in for
specifically this use case, but which turned out to be an incorrect fit.

Note that with the introduction of this package, we have a dependency on
supranational/blst, which is a C dependency; and there is currently the
outstanding issue supranational/blst#245 which prevents building blst
with Go 1.24. However, supranational/blst#247 is an open PR that fixes
the build for Go 1.24 and Go tip.