refactor: extract KeyIDChecker type
Although the two primary implementations of
CommonMessageSignatureProofScheme will rely on counts, which are simple,
this interface still ensures we only allocate for the public key slice
once. It is reasonable to assume that we are eventually going to have a
scheme that does not encode key IDs as positional indices, so I expect
this type to be useful going forward too.