chore: wire previous validator set in round lifecycle
Due to the order of operations when proposing and signing a header, the
finalized previous commit proof is going to be created in the state
machine. The state machine needs knowledge of exactly which validators
were present for those signatures, and prior to this, there was not a
single place to discover that information.