wip: update signature of ValidateFinalizedProof
The sum of the signatures is meaningless if not everyone voted for the
same block. So instead, split the signature bits out by voted hash --
which allows the caller to sum up the total vote power -- and for
convenience, indicate here whether there were any duplicate signature
bits. It is the caller's responsibility to handle double signs.