feat: use finalized proofs in headers
In further preparation of integrating BLS signatures, the state machine
uses a "finalized" commit proof when publishing a proposed block, and
the mirror now interprets and validates that finalized proof properly.

The format of the existing simple common message signature proofs is
effectively unchanged, but BLS signatures will undergo a final
aggregation that is optimized for minimal storage, but which is not
compatible with live consensus gossip.

There is one particular downside to this approach, which is that the
finalized commit proofs can only be backfilled into previous commit
proofs for some signature proof schemes -- as indicated by the
CanMergeFinalizedProofs method.

The proofs that can be backfilled give some allowance on missed
precommit proofs. To provide a similar allowance on aggregated proofs,
we may need to update the storage layer. We can get the proofs through
the stored proposed headers, which would suffice for validating that
everything was signed correctly, but that would not suffice for the
scenario where:

- Validator misses some precommit signatures at height H
- Validator sees proposed header at height H+1, and it has a finalized
  commit proof which proves that height H is committing
- But the proposed header fails to reach consensus for some reason
- On the next round, it is Validator's turn to propose

The state machine would need to know to use the finalized commit proof
on disk instead of consulting its own commit proofs. The proof scheme
could possibly attempt to decompose the finalized proof, which is very
feasible with the right set of commit proofs, but the complexity there
doesn't seem worth it.