wip: handle validating finalized proof with only main signature
The main signature is relatively simple -- next is taking the "rest" of
the signatures while reducing the population of the "n" validators in
our combination indexing. While that isn't terribly complicated either,
it will involve some subtle math that may not be totally straightforward
to follow. We'll keep it as clean as possible.