feat: add finalized format of BLS signatures
In contrast with the signature details transmitted during gossip,
finalized signatures represent a format that is not intended to be
modified any further. With that assumption, we are able to pack
signatures into the finalized format with nearly minimal space. See the
godoc for the SignatureProofScheme for extensive detail.

The one large remaining detail is updating the tmstore package to
account for the finalized signatures not being able to be backfilled in
the way that the simple scheme can. And we probably need separate
special handling for storing the finalized signatures at all, too.