test: simplify creating a new fixture
The unexported fields were a bit of a pain for creating fixtures outside
of the tmconsensustest package -- which came up when creating a fixture
for BLS signatures to use in tmsqlite. So, extract a new function to set
the unexported fields, allowing easy creation of a valid fixture from
outside of tmconsensustest.