fix: load genesis validators if starting at height 3
Height 3 was a particular special case -- a combination of some
off-by-one checks -- that was failing a particular gcosmos test. The
issue there was that we populate a minimal genesis at startup, but it
gets fully populated during InitChain, and then finalization details are
stored. But, if we are on a new run that hasn't gone through InitChain,
and if we happen to be on height 3, then it would inspect an
underpopulated genesis instead of consulting the FinalizationStore.

Now, properly consult the FinalizationStore when possible.