chore: adjust network testing API
This used to accept a context and a logger; now make it accept a
testing.T and a context instead.

The rationale here is that it is easy for a test to provides its own
logger related to the T, but relying on context cancellation for cleanup
is inferior to using T.Cleanup.

The downside of this is that there is no way for a test to set
attributes on the logger, but it doesn't look like any tests set any
attributes so far anyway.