feat: add RoundSessionChanges to network view update
These are strong hints from the consensus engine to the gossip strategy
about the state changes of network sessions. The existing libp2p pubsub
was effectively stateless, so we didn't need it sooner; but dragon uses
dedicated QUIC streams for vote gossip and block propagation, and the
consensus engine ought to be the source of truth for when information no
longer needs to be transmitted on the network.

There are three states according to the consensus engine: Active,
recommending that sessions are live; Grace, recommending that they are
live but will soon be expired; and Expired, which is a recommendation to
close the session.

There is only light test coverage on this so far. I assume there are
many edge cases we are not correctly handling, but this should be enough
of a foundation to fill in missing tests as they arise.