refactor: load validator set for future rounds in same height
This is currently the only case where we know we already have the
information in memory, which allows us to avoid a store lookup.

This is one step towards actually jumping to a future round. We will
also need the vote power hash if it's a later height; we should have
that accessible if we do have the proposed header for that height or its
predecessor, at least.