feat: add ProposedHeaderInterceptor interface
This is specifically to help simplify gdragon in identifying when the
local state machine is proposing a block, and to allow modifying
annotations on the proposed header.

There will need to be some further modifications to register the "block
data" generated in integration tests, but this establishes the core hook
that the gossip strategies need to encode network-related details into a
proposed header.