test: add tmgossiptest DaisyChain implementations
The existing tmp2p types are not used with (g)dragon, and in fact the
tmp2p types are really only used in tests. This mismatch makes it
difficult to add new tests to increase coverage. Therefore, first we are
adding a new daisy chain gossip strategy here in isolation, and the next
change will involve updating the existing tmintegration tests to use
gossip strategies directly instead of the tmp2p types. That may leave
the tmlibp2p package in an unclear state; libp2p support is not
currently high priority, so we will see how it works out.