test: add gossip-strategy-centered integration tests
These currently just port the existing tmp2p-focused integration tests
to use the gossip-strategy-focused ones. We are leaving the existing
tests in place for the moment; the next step is to add a new integration
application to the gossip strategy suite, exercising block data
propagation (the two existing tests avoided block data, instead choosing
to encode details in the data ID).

Once there is at least one integraiton test exercising block data, we
are likely safe to remove the tmp2p tree as it will only be used in
outdated tests.