test: add BlockDataStore for integration tests
The first two integration tests deliberately avoided using block data,
bu the new random data application intentionally uses block data; and so
it needs a way to tell the gossip strategy what data to broadcast.

The integration tests simply provide the data, now by way of the
tmintegration.BlockDataStore, and it is up to the gossip strategy to
retrieve that data and broadcast it. We keep this block data store
limited to the tmintegration package, because it would be reasonable for
a gossip strategy to require a concrete struct as input, serializing at
the last possible moment.